requirements.txt
numpy
faiss-cpu
streamlit
pandas
openpyxl
ollama

